requirements.txt
aws-cdk-lib==2.14.0
boto3==1.21.13
cdk-dynamo-table-view==0.2.0
pytest==6.2.5
pylint==2.12.2
black==22.1.0